refactor: split client message models by file
