+++ Sources/Models/AssistantStarted.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public struct AssistantStarted: Codable {
+    public let newAssistant: JSONValue
+}
