+++ Sources/Models/CallDeleteFailed.swift
@@ -1,0 +1,3 @@
+import Foundation
+
+public struct CallDeleteFailed: Codable {}
