+++ Sources/Models/CallDeleted.swift
@@ -1,0 +1,3 @@
+import Foundation
+
+public struct CallDeleted: Codable {}
