+++ Sources/Models/ChatCreated.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public struct ChatCreated: Codable {
+    public let chat: JSONValue
+}
