+++ Sources/Models/ChatDeleted.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public struct ChatDeleted: Codable {
+    public let chat: JSONValue
+}
