+++ Sources/Models/LanguageChangeDetected.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public struct LanguageChangeDetected: Codable {
+    public let language: String
+}
