+++ Sources/Models/SessionCreated.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public struct SessionCreated: Codable {
+    public let session: JSONValue
+}
