+++ Sources/Models/SessionDeleted.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public struct SessionDeleted: Codable {
+    public let session: JSONValue
+}
