+++ Sources/Models/SessionUpdated.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public struct SessionUpdated: Codable {
+    public let session: JSONValue
+}
