+++ Sources/Models/ToolCalls.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public struct ToolCalls: Codable {
+    public let toolWithToolCallList: [JSONValue]
+    public let toolCallList: [JSONValue]
+}
