+++ Sources/Models/ToolCallsResult.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public struct ToolCallsResult: Codable {
+    public let toolCallResult: JSONValue
+}
