+++ Sources/Models/TransferUpdate.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public struct TransferUpdate: Codable {
+    public let destination: JSONValue?
+    public let toAssistant: JSONValue?
+    public let fromAssistant: JSONValue?
+    public let toStepRecord: JSONValue?
+    public let fromStepRecord: JSONValue?
+}
