+++ Sources/Models/WorkflowNodeStarted.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public struct WorkflowNodeStarted: Codable {
+    public let node: JSONValue
+}
